refactor: rename jwt helper file, add missing newline
